pyproject.toml
[tool.poetry]
name = "fineladybot"
version = "2.0.0"
description = "A Reddit bot that tracks discussions related to Banbury, UK and cross-posts them to the /r/banbury subreddit"
authors = ["Jon <jmcharter@protonmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
certifi = "2021.5.30"
charset-normalizer = "2.0.4"
click = "8.0.1"
idna = "3.2"
mypy-extensions = "0.4.3"
pathspec = "0.9.0"
platformdirs = "2.3.0"
praw = "7.6.1"
prawcore = "2.3.0"
python-dotenv = "0.19.0"
regex = "2021.8.28"
requests = "2.26.0"
tomli = "1.2.1"
update-checker = "0.18.0"
urllib3 = "1.26.6"
websocket-client = "1.2.1"
pydantic = "^1.10.4"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
pytest = "^7.2.1"
mypy = "^0.991"
types-regex = "^2022.10.31.3"
types-setuptools = "^65.7.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.mypy]
files = ["fineladybot/**/*.py"]